extract functions to a centralized helper
